Add interrupt driven UART and further peripherals
The UART connection class was abstracted into a interrupt driven master/
slave connection. The master class was added to AVR_Lib. The slave class
has to written.

The peripheral component was extended with an accelerometer and led
class. The led class can be used with led groups of any size.